docker/requirements/requirements-fastapi.txt
# FastAPI for creating REST API
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
dagster
python-dotenv
